pyproject.toml
[project]
name = "cynditaylor-com-bot"
version = "0.0.1"
description = "SMS-driven agent that edits the cynditaylor-com static site."
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "openinference-instrumentation-bedrock>=0.1.0",
    "strands-agents>=0.1.0",
    "openinference-instrumentation-strands-agents>=0.1.0",
    "bedrock-agentcore>=0.1.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
